Add more default role definitions
This PR adds default role definitions for the following applications:

- Compliance
- Insights
- Drift
- Vulnerability

Each of these applications needs a base, global read/write permission for
itself, as well as inventory, which is a dependent service which will be
receiving permissions from the upstream service. These permissions are set on
a base role for each application.

**Note:** _we are in the process of migrating all default definitions to another
repository, so this PR may just be left for tracking._