Ensure we are validating the format of the role UUIDs before POST/DELETE
This adds the logic and tests to ensure we return a 400 if any of the UUIDs for
roles provided either in a `POST` or `DELETE` are invalid UUIDs, the
request fails with a 400, and explanation, similar to `principals`:

```
{
  "errors": [
    {
      "detail": "Must be a valid UUID.",
      "source": "roles.0",
      "status": 400
    }
  ]
}
```
For requests where the UUID is valid, but the Role record is not found, we will
ignore the UUID, but will not fail the request. In the future we _may_ want
to consider a 207 in these cases and return feedback of what did/did not get
saved or removed.