Account for unique constraint on policy name when creating system policies
Since there is a null constraint, and unique constraint on the policy model,
we need to supply a unique, and constant name for the policy that we can reliably
use to look the policy up with.

Since the policy is for a group, and the UUID won't change on the group, we
can use that.