Ensure default group roles are inherited by principals of a tenant
This PR ensures that the group access for a principal is the union of groups
that they've been explicitly assigned to, as well as the tenant's `platform_default`
group.

This will cascade to the policy/role/access methods for a principal to ensure
that the access for all users of a tenant includes roles defined as part of
the `platform_default` group.

Since the management of this `platform_default` group will be determined by the
`system` flag, even once an org_admin changes the roles within this group, all
principals will continue to inherit its roles and access.