Abstract prefetch options for querysets
Since we reuse a lot of the same methods for different resources, the previous
commit could have an unintended side-effect of eager loading roles/access when
we don't need to, like for groups scoped by principal for instance, which will
increase the load time.

This abstracts the options for the id sort query, and the group query. We may
be able to improve some of the other endpoints' performance by optionally including
more prefetch options for those scoped to principal as well.