Address n+1 queries in GET `/roles/?scope=principal`
Similar to #172, when roles are scoped to principal, we end up running into
n+1 issues due to needing the policy/role records for the groups for a principal.

This resolves that by added the necessary prefetch lookups on the `get_role_queryset`
calls, and sets up a base query to be reused/extended upon.