Add data migration to remove default seeded roles
In order to not make system and/or platform_default roles available or enforced
for applications not integrating with the GA rollout of RBAC, we need to remove
these roles from the seed files for the following apps:

- ansible-automation
- ansible-hub
- catalog
- remediations
- sources
- subscriptions

Other apps not needing default roles are:

- inventory
- approval
- sources
- cost-management

However at the time, we don't have default roles for those apps, so we don't need
to take any action.

A dependent PR has been added to the [RBAC config repo](https://github.com/RedHatInsights/rbac-config/pull/17)
to remove these roles from the seed files. That will need to be merged and rolled
out before this PR can be merged.