Add ?principal_username= filter on /groups/<uuid>/principals/
This reuses a similar pattern added for filtering roles on groups, so it's now
broken out into a helper to build the filters for nested resources on groups.

Now, a filter `?principal_username=` is supported on /groups/<uuid>/principals/
in order to return principal details for a group, filtered by username.