Ensure username_list includes all principal data
Previously, `username_list` was being populated with just the first value in
the list from:

```
list(usernames[0].values())
```

Since `usernames` in this case was a list of dictionaries, it only grabbed the
value of the first entry, which results in the `/groups/<uuid>/principals/`
endpoint only ever returning one principal.

This change ensures that we're building a list from all items in the serialized
data.