Add primary and alt-key check, ensuring identity is honored before service auth
We need to ensure that if an identity header is supplied, that it takes precedence
over PSK/service auth. This way, we ensure that any traffic coming through c.r.c
will use the identity provided, only allowing for PSK/service auth to be used
inside the cluster.

Similar to the patterns in the gateway and BOP, we should allow for key rotation
per service, by respecting a primary and alternate key.