Ensure /role/?username= returns count data
A bug was discovered where `accessCount` and `policyCount` data was missing from
the dataset when `?username=` was supplied on the `/role/` GET endpoint.

This is because of the fact that the queryset is different based on whether or
not that username param is supplied. The annotations for those two fields need
to be added.