Ensure the default group query for a tenant returns the system groups
If the default group has been modified, meaning it's no longer a `platform_default`
group, the roles associated with the default group aren't included in the filtered
results from `/roles/?username=<someuser>`

This issue is due to the fact that when a default group is updated and thus is no
longer managed by the platform, we flip the `platform_default` flag on the group
to `False`. The `system` flag is still `True`, which is how we know which group
for a tenant is that "default" group that all principals inherit roles from. The
platform_default informs us of whether or not it's been updated and is managed by
the platform.

Currently, our query on the group model to retrieve the platform default queryset [1]
filters based on the `platform_default` flag. This means that after a default group
has been modified, that query will no longer return the roles for that group when
used to retrieve a user's roles.

The fix for this should just be to change the query flag to use `system`.

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/group/model.py#L55