Ensure the /access/?application= check is on permission name only
There's currently a bug where the `/?application=` param value is being compared
against the entire permission string. This is causing scenarios where resource types
or actions in the permission can cause it to match on the application name since
it's a partial match on the entire permission.

Here is [an issue](https://github.com/RedHatInsights/insights-rbac/issues/264) that was logged against this as well.

For instance, a permission `catalog:*:foobar` would be returned with a query:
`/access/?application=foo` when it should only respect the `catalog` substring
as the partial being matched against.

Also, the same permission would return for a query: `/access/?application=cat`.

This does two things:
- restricts the matching to the application name of the permission only
- requires an exact match, to align with the spec.

**Testing Locally**
- checkout master
- modify a permission on a role that would normally return on a given ?application=
  query so that the it looks something like `app-name:foo:bar`
- query `/access/?application=foo` and see that the permission returns
- query `/access/?application=app` and see that the permission returns
- now checkout this feature branch and run the same query against `foo` and `app`.
  you should not see the permission returned, but `/access/?application=app-name`
  should work.