Fix tenant creation exception
Can't create tenant outside the public schema is thrown, which is caused
concurrent requests: when a bunch of requests try to get tenant, it does
not exist, so they try to create one but only one succeeded, others throw
exceptions not handled.

This PR tries to fix that issue by using get_or_create instead of create.