Ensure /roles/?add_fields=groups_in is scoped correctly
When `/roles/?add_fields=groups_in` or `/roles/?add_fields=groups_incount` is accessed
with `?username=` or `?scope=principal`, we should ensure that the data in the
`groups_incount` and `groups_in` payload is scoped to the user, either by username
or the principal making the request, depending on which parameter is passed in.

This does that by grabbing the principal from the request (username or scope) if
either of those params exist, and filtering the group query to the principal in
question.

TODO:
- add specs to exploit this
- update the openapi.json spec with a note about the behavior