Ensure we close thread connections while seeding
We need to ensure we close the DB connection(s) which end up being opened with
`ThreadPoolExecutor()` explicitly, otherwise they are not released, and we end
up hitting our PostgreSQL limit in environments where this is set to the default (100).

What happens is that when we run seeds, each tenant/thread opens a new database
connection. When the number of tenants exceeds the number of connections, the async
processing can quickly exhaust all available database connections, since they
are not immediately released.

We need to ensure that after each thread is completed, we close the connections for
that thread. This should resolve the issue, but we can also limit this further
by setting an explicit `max_workers` value to limit the concurrent threads being
run for seeding. When this is _not_ supplied, it will use the default value for
the concurrent threads [1].

[1] https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.ThreadPoolExecutor