Fix platform_default group cache invalidation
There are triggers on all data relations which would affect a user's permissions.
This includes adding/removing roles to/from a group. When this happens, the cache
for all users in a group's policy is invalidated.

The problem is that for `platform_default` groups, there is no actual relation
between users and the group/policy/roles for that group. This means it effectively
bypasses the cache invalidation logic [1].

The fix for this is to conditionally invalidate all users' cache when the
`platform_default` group is updated.

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/policy/model.py#L56