Add settings for parallel migrations
This allows for us to configure either in e2e-deploy, or via the command line,
the number of parallel processes used for running migrations, as well as the number
of migrations each worker should process at one time [1].

We'll preserve the default values (2). From the command line, this could be overriden
with something like:
```
TENANT_PARALLEL_MIGRATION_MAX_PROCESSES=20 \
TENANT_PARALLEL_MIGRATION_CHUNKS=5 \
./rbac/manage.py migrate_schemas --executor=parallel
```
It should be noted that in in pre-production environments, we have 100 database
connections. In production we have upwards of 1,000, but should be mindful of this
when running migrations with these values set explicitly.

[1] https://django-tenant-schemas.readthedocs.io/en/latest/use.html#migrate-schemas-in-parallel