Add ?name_match=partial|exact on /groups/ GET
This allows a way to specify whether the `/groups/?name=` param should be a partial
(default) or exact match.

When `name_match` is not supplied in the request, the query will default to the
current behavior of being a partial match. There are two valid values for this
param, `partial` and `exact`.

If `partial` is supplied, it will behave the same, just explicitly so. If `exact`
is provided, the match will be an exact match, and only return the group with
the exact name provided, if it exists.

This will allow for the UI to validate group names on creation, prior to the
actual POST to create the object.

If an invalid option is supplied, a 400 will be returned with a message informing
the client of the valid values.