Factor out CommonFilters
This sets up a `CommonFilters` class which inherits from `filters.FilterSet`, which
the model filters can then inherit from, to get the default behavior as well as
any common filters we may want to add.

This initially consists of the name filter, which is generalized to be used against
any of our models with name fields.