Setup initial model/view/serializer for /permissions/
The UI will need a way to retrieve all `Permission` records for a tenant, and
will also need a way to filter these results by `application`, `resource`, and
`operation`.

This sets that up by factoring out the `Permission` model, adding the url routes,
a view for GETting a list of permissions, and a serializer to return the paginated
results.