Fix role seed Access creation for new roles
The existing logic for setting `version_diff` during role creation [1] causes
`version_diff = defaults["version"] != role.version` to always evaluate to `False`
because we're using `defaults["version"]` to create the new role with, causing
`role.version` to be the same.

This means that when creating a new role, the `version_diff` check here [2] which
used to force `Access` objects to be created, will not be invoked.

This updates the check when setting `version_diff` to maintain the previously
functional behavior by ensuring a truthy value when a new role is being created.

Previously the expectation was that `role.version` would be a null value, thus
being able to use that in the check for `Access` creation since the value from the
config file would be a valid number value. This ensures the same behavior.

Alternately, we could change this check [3] to explicitly check against `created`,
and leave line 43 as is.

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/role/definer.py#L43
[2] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/role/definer.py#L55
[3] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/role/definer.py#L55