Remove `update` flag in role seeds to simplify logic and fix bug
Submitting this in favor of #312 in order to simplify the logic around role
seeding when we're updating vs creating.

This will ensure that when we're creating a new role, we are creating access
records from permissions. When we're updating and the version differs, we'll
remove the access records and recreate them. When we're updating and the version
is _not_ different, we'll update the role, but return early and not do anything
with the access records.

This also fixes the original bug in #312 while reducing complexity.