Catch and log role/permission seeding errors
Currently, since seeds are executed in threads, if there's an error creating or
updating a role or permission from seeds, it will fail silently. The next tenant
will continue to seed, but the iteration of seed/permissions for the tenant with
a failure will stop.

This ensures that we log any errors, and continue the seeding for the tenant.