Ensure the "applications" field is set on /api/rbac/v1/roles/<uuid>/
If I visit `/api/rbac/v1/roles/<uuid>/` I should see the `applications` field
populated with the same value(s) as I see for the role when viewed at `/api/rbac/v1/roles/`.

Currently it is set as `null` because we don't return the value.

For instance, if I see the following for role `1234` on `/api/rbac/v1/roles/`:
```
{
  "data": [
    {
      "uuid": "1234",
      "name": "test role",
      "display_name": "test role",
      "description": "A description of test role",
      "created": "2020-08-26T01:33:19.639405Z",
      "modified": "2020-08-26T01:33:19.647813Z",
      "policyCount": 0,
      "accessCount": 1,
      "applications": [
        "cost-management"
      ],
      "system": false,
      "platform_default": false
    }
  ]
}
```

I should see the following on `/api/rbac/v1/roles/1234/` (note the `applications` field):
```
{
  "uuid": "1234",
  "name": "test role",
  "display_name": "test role",
  "description": "A description of test role",
  "access": [
    {
      "resourceDefinitions": [],
      "permission": "cost-management:*:read"
    }
  ],
  "policyCount": 0,
  "accessCount": 1,
  "applications": [
    "cost-management"
  ],
  "system": false,
  "platform_default": false,
  "created": "2020-08-26T01:33:19.639405Z",
  "modified": "2020-08-26T01:33:19.647813Z"
}
```