Add initial restricted internal API access
This change will introduce a prefix `/_private/` into the RBAC API, which will
be gated by the new `InternalIdentityHeaderMiddleware`, and will require that
the identity header be one from Turnpike.

We will need to setup the config in Turnpike similar to:
```
- name: rbac
  route: /api/rbac
  origin: http://rbac.svc.cluster.local:8080/_private
  ...
  ...
```

which will expose the `_private` routes to internal users. Anything coming through
the 3scale gateway on https://cloud.redhat.com/api/rbac/v1/* will continue to be
served through the current middlewares, and any attempt to request `_private`
through 3scale or directly will fail with a 403.

We will also short-circuit the current `IdentityHeaderMiddleware` during `process_request`
and `process_response` if the path matches a private URL, and ensure the `InternalIdentityHeaderMiddleware`
check is run (which will no-op `/api/` paths).

In addition to this setup, we've added two private routes:
```
/_private/api/tenant/unmodified/
/_private/api/tenant/<str:tenant_schema_name>/
```

The former will return the `schema_name` of all tenants which meet the following
criteria, deeming them "unmodified":

- tenant only has one group
- that one group is a system group (meaning the platform default that has _not_ been modified)
- tenant only has system roles (meaning they haven't created any new roles)
- the total number of groups == 1
- the total number of roles == the number of system roles

The latter will allow for a system endpoint in order to delete a schema. This will
ultimately help to prune tenants which haven't done anything with their RBAC
data, and thus can afford to be deleted and recreated if they access RBAC again,
helping to clean the ~30k schemas we currently manage in production now that
[this ticket](https://issues.redhat.com/browse/RHCLOUD-8796) has been fulfilled.