Add /_private/api/migrations/progress/?migration_name=<name>&app=<app>
This will allow us to get the status of a particular migration across all tenants.
By default the `app` will be `management`, but you can optionally supply the `app`
param for checking against `api` or other Django apps' migrations.

Given that we know what the last applied migration should be for an app during
a release, we can use this to supply that migration name to get back the completion
progress during a migration run, which will look like:

```
http://localhost:8000/_private/api/migrations/progress/?migration_name=0004_crossaccountrequest&app=api
```

```
{
  "migration_name": "0004_crossaccountrequest",
  "tenants_completed": 3,
  "total_tenants_count": 6,
  "percent_completed": 50
}
```