Add internal endpoint for triggering seeds
Adds a `POST` endpoint at `/_private/api/seeds/run/` with optional params for
specifying which seeds to run with `?seed_types=roles,permissions,groups`.

When this is not specified, it will default to all, similar to the underlying
Django command.

If an unsupported seed type is presented, it will 400 the request in order to
help prevent seeds from failing silently in a background process.