Verify permission exists on role creation
We currently check against our allow-list to ensure permissions of a role being
created have their `application` field defined in list.

However, iff that check passes, you are still able to create a permission for any
`resource_type` or `verb`, which may or may not be a valid permission. For example,
given the following allow-list:

```
ROLE_CREATE_ALLOW_LIST = "cost-management"
```

This role creation would succeed:
```
{"name": "Custom Role", "access": [{"permission": "cost-management:foo:bar", "resourceDefinitions": []}]}
```

And this role creation would fail:
```
{"name": "Custom Role", "access": [{"permission": "rbac:foo:bar", "resourceDefinitions": []}]}
```

Assuming we don't have a permission defined as `cost-management:foo:bar`, this
would not be what we want. The changes in this PR will enforce the `resource_type`
and `verb` check, to ensure a permission in the DB defined with `cost-management:foo:bar`
exists first.

To test this locally:

- on `master`, try to create a role with a permission for an app which exists in
  your local `ROLE_CREATE_ALLOW_LIST`, with a random `resource_type` and `verb`
- note that this currently works
- checkout this feature branch
- run the same POST as you did on `master`
- this should now fail with a 400 and helpful message
- now send another POST, this time with a permission string you _do_ have locally
  in your db
- this should create the role