Add the `management` models to the public schema
To support the upcoming TAM cross-request access features as well as moving away
from Django Tenant Schemas, this will sync the `management` app models with the
public schema structure moving forward.

Initially, we'll be seeding the `system` data in this schema in order to populate
the front-end with the roles and permissions available for cross-access requests,
which will ultimately be tied via a many-to-many relation to cross-access request
object.

Eventually, these models will be kept in sync with individual tenant schema
data, to migrate away from a multi-schema tenant architecture.