Output the permissions which would be deleted in seeding
In order to ensure we don't delete any orphaned, valid permissions in any environment,
we should first run this change with logging (non-destructive) to output these
potential permissions.

If there are any, we'll need to verify with the app team the permission is for that
it should be added to their permission config file.

Once added, we can safely run the seeds and these permissions won't be deleted from
the database.