Output the roles which would be deleted in seeding
In order to ensure we don't delete any orphaned, valid roles in any environment,
we should first run this change with logging (non-destructive) to output these
potential roles.

If there are any, we'll need to verify with the app team the role is for that
it should be added to their role config file.

Once added, we can safely run the seeds and these roles won't be deleted from
the database.

These are filtered to only system roles, so we don't delete custom roles.