Add TenantAwareModel
This adds a `TenantAwareModel` class, which inherits from `model.Model`. This allows
us to reuse this class on all models which will have a relation to the `Tenant`
model, meaning we will just need to have those models inherit from `TenantAwareModel`
to pick up any behavior related to the tenant relation.