Override `create` method in `ModelSerializer`
This overrides the `create` method in the `ModelSerializer` to ensure that we are
able to set the `tenant_id` for any object created in the API.

We'll still need to handle seeds by passing in the tenant, and any other object
creation that happens outside of views. In the future, we'll want to enforce this
via RLS.

We cannot add the `tenant_id` in a signal, because we don't have access to the
request, where the tenant exists. We also cannot do this directly in the views,
since it would require we expose the `tenant_id` field in the serializers which
we don't want to do.

Instead, this provides us the opportunity to add the `tenant_id` after serialization
has succeeded, but before the object is saved. This is abstracted out so we can
include it as a mixin to the serializer classes which inherit from `ModelSerializer`.

The existing behavior is copied directly from our current version 3.12.2 https://github.com/encode/django-rest-framework/blob/3.12.2/rest_framework/serializers.py#L904
and injects the `tenant_id` in the `try` block.