Ensure `tenant_id` is saved on all creates
This makes use of the `SerializerCreateOverrideMixin` in order to add the behavior
of saving the `tenant_id` from the request object on creates.

There are also some instances where serializers implement their own create, or
are not model serializers. In these cases, and other cases where we're doing an
explicit `create()` or `save()` we also need to ensure the `tenant_id` is set.

For `get_or_create()` calls, currently we don't have `tenant_id` set on all objects.
If we add `tenant=tenant` directly into these calls, we'd create duplicate records
since the `get` would not match on the tenant identifier. To get around this, we'll
need to update the object after `get_or_create()` to add the `tenant_id` value if
it does not yet exist.

We should be able to pull these noted references out after we're strictly enforcing
the FK constraints and have confirmed all records have `tenant_id` values. At that
point we should be able to add `tenant=tenant` to the `get_or_create()` directly.

One other consideration is that this commit may be something we'll want to revert
if/when we have RLS (row-level security) in place. If we do it correctly, we would
ideally have RLS policies setup for all operations (SELECT, INSERT, etc), which would
mean we'd no longer need to update all queries directly with the tenant, as the
policy would set the `tenant_id` in the middleware.