Ensure Role objects have `modified` updated when seeds change
When a `Role` object is updated via seeds and the defaults on the actual object
change, the `modified` field will update via Django. The same is true if `Access`
objects are added/removed to/from a `Role` via the API.

However, when seeding a change to `Access` (permissions) only on a role, and the
version number is bumped, we update the `Role` with defaults and in this case it
would have no diff, so Django would not update the object, thus not updating the
`modified` field. We then remove and add the `Access` objects outside the context
of an API resource and update the model directly, which does not update the `Role`
either.

This explicitly sets the `modified` field to `timezone.now()` which is the default
on the model, whenever the version changes in the seed files. This will ensure it
updates regardless of whether or not the `Role` data itself is updated.