Copy any "custom" permissions per tenant
There are cases in prod where a permission was removed prior to us seeding the
public schema. This was actually a "rename", where any custom default groups
using the role with this permission would have held onto the old permission.

In order for all roles to copy with the correct permissions, we'll need to sync
any permissions for a tenant which don't exist in the public schema.