Add migration for dependent permissions
Moving this migration up ahead of the model/code changes to ensure compatibility
once the code is deployed.

The migration is based on the following `Permission` model change:
```
permissions = models.ManyToManyField("self", symmetrical=False, related_name="requiring_permissions")
```

This will add a table `management_permission_permissions`, so that there can be
a many-to-many relation between permissions. `symmetrical=False` will ensure
that a many-to-many `self` reference will not create a symmetrical entry when
addinig dependent permissions to a record [1].

[1] https://docs.djangoproject.com/en/3.2/ref/models/fields/#django.db.models.ManyToManyField.symmetrical