Relate permissions during seeding
Since permissions cannot be created in the API, we'll be relating them when we
seed permission data.

We cannot do this as part of the initial permission seed creation, because of the
fact that there may be a dependency on a permission which has not yet been created.

Instead, we'll map out any permission seed data which has `requires` after we've
created the permissions, and make the associations at that point.

There were some issues with catching errors in `transaction.atomic()` [1] so this
moves the initial transaction block and adds another.

[1] https://stackoverflow.com/questions/42586967/using-nested-django-transaction-atomic-to-handle-exceptions