RBAC on RBAC for /principals/ reads
*Jira:* https://issues.redhat.com/browse/RHCLOUD-16704

This adds support for a new system role in RBAC which will allow for read-only
access against the `/principals/` endpoint, when granted.

Previously, this endpoint was restricted to admins only. We've added a new permission
class `PrincipalAccessPermission` which checks for the same `ALLOW_ANY` or `request.user.admin`
allowances as the `AdminAccessPermission` class does, and also adds a check for
a permission: `rbac:principal:read` on `GET` requests.

When the user (even non-admins) have this access granted to them, they'll have
read-only access to principal data for their org.